convert poll interface to use "future" pseudo-handles
Per discussion with pchickey, this modifies the poll API to use "futures" rather
than a `subscription` type.  The former is arguably closer to the API we'll use
once resource support is finalized.

You can create a future by calling `subscribe-read`, `subscribe-write`,
`subscribe-wall-clock`, or `subscribe-monotonic-clock` and then poll any number
of futures using `poll-oneoff`.  The result is a list of booleans indicating
which futures have completed.

In order to determine how many bytes are readable or writable, the application
may call `bytes-readable` or `bytes-writable`, respectively, passing the
descriptor of interest.

Finally, a future pseudo-handle may be disposed of using `drop-future`,
analogous to disposing of a resource handle.

Note that the updated API does not include `userdata` type, fields, or
parameters, so the Preview 1 adapter will need to track them itself using a
table of some kind.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>