sketch of `poll_oneoff` polyfill
This adds a `poll_oneoff` implementation to src/lib.rs.  It's completely
untested.  I was planning on adding a host implementation and using that to test
end-to-end, but that raised some tough questions about how much of the existing
`wasi-common` scheduler(s) should be reused.  I've decided to focus on other,
more widely-used parts of WASI first, but I wanted to share the work I've
already done here.

Note that I've moved the clock- and socket-specific functions out of `wasi-poll`
and into `wasi-clocks` and `wasi-tcp`, respectively.  The latter is a new
interface which will eventually contain functions and types resembling
@npmccallum's https://github.com/npmccallum/wasi-snapshot-preview2#wasi-tcp
proposal.

Per discussion with @sunfishcode:

- `wasi-tcp` includes an `error` enum type, intended to represent only socket-related errors.
- It also includes a `socket` pseudo-handle type, distinct from `wasi-filesystem`'s `descriptor` type.

These fine-grained types help move us away from the "everything is a file
descriptor" and "all errors are errnos" approaches of Preview 1.

If we decide `poll-oneoff` should be usable with files as well as sockets, we
can add `subscribe-read` and `subscribe-write` functions to `wasi-filesystem`
which accept file `descriptor`s.  Likewise for any other pseudo-handle type from
which we'd like to create futures.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>