Upload a release artifact from CI
This updates CI to upload a `wasi_snapshot_preview1.wasm` release
artifact from CI. One is stored per CI job in case it needs to be
inspected, and additionally a github action is used to maintain a
"latest" release for pushes to `main` to ensure that the latest copy of
`wasi_snapshot_preview1.wasm` is available.

I haven't thought much about official releases or tags or things like
that but figured that this was a good starting place at least.