adapt `wasi-common` for use in host implementation
Per recent discussion with Dan and Pat, this begins the work of adapting
`wasi-common` for use in the Preview 2 host implementation.  It removes the
`wiggle` dependency and the rights system, and adds a few structs and fields
specific to Preview 2.  Further changes will follow as we implement more
features.

The intention is that this version of `wasi-common` will be merged back into
`wasmtime` and be used there for both the Preview 1 and Preview 2 WASI host
implementations.  In addition, we would like it to remain runtime-agnostic
(i.e. not depend on `wasmtime` directly or indirectly) so it can be reused by
other runtimes, which was the original intention behind `wasi-common`.
Currently, this requires a manually-written conversion from `wasi_common::Errno`
to `wit_bindgen_host_wasmtime_rust::Errno`; there might be a way to automate
that conversion as part of code generation later on.

This also enables `async` code generation, since most of the `wasi-common`
methods are `async` (though none of them currently `await` anywhere, and many
block).  The intention is that only `wasi-poll::poll-oneoff` will actually
`await`, and we'll replace the current schedulers in `wasi-common` with
something based on e.g. `tokio`.

Other minor changes:

- `wasi-clocks::wall-clock-resolution` now returns an `instant` instead of a
  `datetime` (I presume the latter was a mistake).
- The structs in `host::clocks` have been moved to `wasi-common::clocks`.
- `host::WasiCtx` has been replaced by `wasi-common::WasiCtx`.

Finally, this implements `wasi-filesystem::pread` and `wasi-filesystem::pwrite`
using `wasi-common::WasiFile`.  Implementing the rest of the interface should be
straightforward.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>