Teach poll_oneoff how to poll for immediately-available I/O
When a `ReadPipe` is reading from a memory buffer, there is no host file
descriptor to poll with, but there doesn't need to be because the
contents are immediately available. Teach the `poll_oneoff` function to
use system-interface's `ReadReady` trait, which returns the number of
bytes read to be read immediately, to implement polling for such input
sources.

This also fixes the polyfill to avoid calling `bytes_readable` on
non-socket handles, which is needed to make the test pass.