short-circuit reentrance when allocating stack and `State`
Per https://github.com/bytecodealliance/wasm-tools/pull/919,
`wit-component` needs to lazily allocate the adapter stack to avoid
premature or infinite reentrance from the main module to the adapter.
This means adding an additional `allocation_state` global variable and
using it to determine when to return early from an exported function,
e.g. because we're in the process of either allocating the stack or
allocating `State`.

This requires an updated `wit-component` dependency once
https://github.com/bytecodealliance/wasm-tools/pull/919 has been
merged.

Fixes #78

Signed-off-by: Joel Dice <joel.dice@fermyon.com>