Make `read-via-stream` etc. infallible.
Wrap `File`'s contents in an `Arc` so that file handles and stream
handles can share the underlying file descriptor, so that we don't have
to call `try_clone()`. This avoids the possibility of `try_clone()`
failing.

This also makes `File` similar to `Dir`, which already has an `Arc`, in
preparation for merging `File` and `Dir` now that WASI has decided that
file and directory shouldn't be separate types.