Add the proxy world, and remove HTTP from command.
Add the proxy world, from the wasi-http repository, and add a
`--world` flag to the `host` command for selecting which world to run
them in. Current options are `command` and `proxy`.

Also, remove HTTP APIs from the command world, because command is meant
for CLI commands that run with a `main` function with command-line
arguments, so it doesn't include programs that run with incoming HTTP
requests

This also removes HTTP APIs from the reactor world. This world doesn't
yet have a standards proposal and doesn't yet have a clear purpose, but
it's otherwise currently in sync with the command world, and this PR
restores it to that.