wit: remove resource index from monotonic-clocks and wall-clocks, delete instance-*-clock interfaces
At this time, we don't have use cases for providing multiple wall clocks
or multiple monotonic clocks to the same component. So, according to the
principle of following preview 1's design as close as possible, we are
changing these interfaces to being provided by ambient functions, rather
than methods on a resource.