Simplify the sockets implementation.
Use the new cap-net-ext crate to simplify the sockets implementation.
Instead of the `TcpSockImpl` enum and tracking the state of
socket->bind->listen->accept explicitly, these are now separate host
calls, which gives us better fidelity to the WASI API, and we don't
need to track this state ourselves.

See wasi-common/cap-std-sync/src/net.rs for the main changes.