filesystem wit: delete non-blocking from descriptor-flags, and delete set-flags
non-blocking or blocking is no longer a property of a file, it is a
choice the caller makes when calling the Input/OutputStream read/write
vs read_blocking/write_blocking method.

Therefore, do not make the non-blocking flag something the user can
specify, modify, or observe about a file.

The set-flags method was only valid for changing the non-blocking flag,
so get rid of it altogether. Target operating systems don't support
modifying any other flags of open files.