Rename wasi-base to wasi-cli-base.
This is just a rename, and it's not yet clear what shape wasi-cli-base
will ultimately take, but for now, it's useful to avoid confusion about
whether wasi-base is meant to be used by all WASI worlds. It is a base
for CLI-like worlds that have command-line args, environment variables,
and exit statuses, and not all WASI worlds will have those.